Revert "Feature/#10 workouts"
