--- jest.config.json
@@ -1,8 +1,0 @@
-{
-    "reporters": [
-        "default",
-        ["./node_modules/jest-html-reporter", {
-            "pageTitle": "Test Report"
-        }]
-    ]
-}
