+++ models/user.js
@@ -4,16 +4,11 @@ const exerciseSchema = mongoose.Schema({
     name: { type: String, require: true }
 });
 
-const workoutSchema = mongoose.Schema({
-    exerciseId: { type: mongoose.Types.ObjectId, require: true },
-    content: { type: String, default: '' }
-});
-
 const sessionSchema = mongoose.Schema({
     name: { type: String, require: true },    
     date: { type: Date, require: true },
     location: { type: String, default: '' },
-    workouts: { type: [workoutSchema], default: []}
+    workouts: { type: [], default: []}
 });
 
 const userSchema = mongoose.Schema({
@@ -24,6 +19,5 @@ const userSchema = mongoose.Schema({
 });
 
 module.exports.Exercise = mongoose.model('Exercise', exerciseSchema);
-module.exports.Workout = mongoose.model('Workout', workoutSchema);
 module.exports.Session = mongoose.model('Session', sessionSchema);
 module.exports.User = mongoose.model('User', userSchema);
