+++ test/authentication.spec.js
@@ -13,7 +13,7 @@ afterAll(async () => await dbHandler.closeDatabase());
 describe('POST /auth/signup', () => {
 
     describe('given a valid signup request with username and password', () => {
-        it('should return a 201 - Created response and create a user record on the database', async () => {
+        test('should return a 201 - Created response and create a user record on the database', async () => {
             const requestBody = {
                 username: 'username',
                 password: 'password'
@@ -32,7 +32,7 @@ describe('POST /auth/signup', () => {
     })
 
     describe('given a signup request with missing username or password', () => {
-        it('should return a 400 - Bad Request response with an error message (missing parameters) and create no records', async () => {
+        test('should return a 400 - Bad Request response and create no records', async () => {
             const expectedMessage = { message: `Signup: Please provide a valid username and password.` };
 
             const noUsernameResponse = await supertest(app).post('/api/auth/signup').send({ password: 'password' });
@@ -49,55 +49,8 @@ describe('POST /auth/signup', () => {
         });
     });
 
-    describe('given a signup request with usernames too long and too short', () => {
-        it('should return a 400 - Bad Request response with an error message (username length) and create no records', async () => {
-            const expectedResponse = { message: `Signup: Username length must be between 4 and 20.` };
-
-            const responseUsernameTooShort = await supertest(app)
-                .post('/api/auth/signup')
-                .send({
-                    username: 'xx',
-                    password: 'password'
-                });
-            expect(responseUsernameTooShort.statusCode).toBe(400);
-            expect(JSON.parse(responseUsernameTooShort.text).message).toEqual(expectedResponse.message);
-
-            const responseUsernameTooLong = await supertest(app)
-                .post('/api/auth/signup')
-                .send({
-                    username: 'xxxxx_xxxxx_xxxxx_xxxxx',
-                    password: 'password'
-                });
-            expect(responseUsernameTooLong.statusCode).toBe(400);
-            expect(JSON.parse(responseUsernameTooLong.text).message).toEqual(expectedResponse.message);
-
-            User.find().then((users) => {
-                expect(users.length).toBe(0);
-            });
-        });
-    });
-
-    describe('given a username containing invalid character(s)', () => {
-        it('should return a 400 - Bad Request response with an error message (invalid characters) and create no records', async () => {
-            const expectedResponse = {message: `Signup: Username can only contain alphanumerical characters and underscore.` };
-            
-            const response = await supertest(app)
-                .post('/api/auth/signup')
-                .send({
-                    username: 'xxxxx!$%#@/',
-                    password: 'password'
-                });
-            expect(response.statusCode).toBe(400);
-            expect(JSON.parse(response.text).message).toEqual(expectedResponse.message);
-
-            User.find().then((users) => {
-                expect(users.length).toBe(0);
-            });
-        });
-    });
-
     describe('given a signup request with a username that already exists', () => {
-        it('should return a 409 - Conflict response and create no records', async () => {
+        test('should receive a 409 - Conflict response and create no records', async () => {
             const username = 'username';
             const password = 'password';
 
@@ -125,12 +78,10 @@ describe('POST /auth/signup', () => {
 
 describe('POST /auth/login', () => {
     describe('given a login request with an existing username and a matching password', () => {
-
-        it('should return a 200 - OK response along with a signed JWT Token', async () => {
+        test('should return a 200 - OK response along with a signed JWT Token', async () => {
             process.env.JWT_SECRET = 'temp-key-for-testing';
-
             const expectedMessage = { message: 'Login: Authentication successful.' };
-
+            
             const username = 'username';
             const password = 'password';
             const hash = await hashPassword(password);
@@ -153,9 +104,9 @@ describe('POST /auth/login', () => {
     });
 
     describe('given a login request with an existing username but an invalid password', () => {
-        it('should return a 401 - Authentication failed with no token attached to the response', async () => {
+        test('should return a 401 - Authentication failed with no token attached to the response', async () => {
             const expectedMessage = { message: 'Login: Authentication failed.' };
-
+            
             const username = 'username';
             const password = 'password';
             const invalidPassword = 'invalid-password'
@@ -179,7 +130,7 @@ describe('POST /auth/login', () => {
     });
 
     describe('given a login request with missing username or password', () => {
-        it('should return a 400 - Bad Request with no token attached to the response', async () => {
+        test('should return a 400 - Bad Request with no token attached to the response', async () => {
             const expectedMessage = { message: 'Login: Please provide a valid username and password.' };
 
             const noUsernameResponse = await supertest(app).post('/api/auth/login').send({ password: 'password' });
@@ -197,7 +148,7 @@ describe('POST /auth/login', () => {
     });
 
     describe('given a login request with a username that does not exist on the database', () => {
-        it('should receive a 404 - Not found with no token attached to the response', async () => {
+        test('should receive a 404 - Not found with no token attached to the response', async () => {
             const username = 'username';
             const password = 'password';
